Edited classes in index.html and catalog.html
Added the regular site-list__item class to the element with the site-list__item--current class in index.html. In catalog.html changed the address according to the layout. In form.html, added the href attribute with the corresponding address to the links leading to the main page, added the site-list__item--current class to the link of the current page in the main navigation and deleted href attribute